fts: extract reader/cursor.rs (TermMeta, BlockMeta, TermCursor)
Phase 1a of the read-path split. Move the low-level posting-cursor
layer — the parsed per-term header/skip table (TermMeta, BlockMeta) and
the block-at-a-time TermCursor — out of reader/core.rs into
reader/cursor.rs. The scorers, phrase walk, and count kernels still in
core read these types' fields and call their methods directly, so the
fields/methods are widened to pub(super) (scoped to reader/, never past
the FTS layer); TermCursor stays pub(crate) because PreparedClauses
carries it. read_u32_le/read_u64_le stay in core (used across it) and
are widened to pub(super) for the cursor parser. Pure move — no logic
change; FTS oracle/count/query suites green.
